+++ changelog.md
@@ -1,5 +1,10 @@
 ## SpellActivationOverlay Changelog
 
+#### v0.9.4 (2023-09-xx)
+
+- New GAB: Mage's Ice Lance button glows during Fingers of Frost
+- New GAB: Mage's Deep Freeze button glows during Fingers of Frost
+
 #### v0.9.3 (2023-09-02)
 
 - Bump in TOC file for Classic Hardcore patch